[c#] Rename Bond.BondReflection back to Reflection
In commit 9753e3be3, the type `Bond.Reflection` was renamed to
`Bond.BondReflection`. This inadvertently broke source- and
binary-compability.

`Bond.BondReflection` has been renamed back to `Bond.Reflection`. To
minimize breakage for anyone who took a dependency on the
`Bond.BondReflection` type in the interim, `Bond.BondReflection` is now
a shim that forwards to `Bond.Reflection`. The methods in
`Bond.BondReflection` are marked as obsolete (warning level,
suppressible).

The internal type `Bond.Reflection.Reflection` in `Bond.Reflection.dll`
has been renamed to `Bond.Internal.Reflection.ReflectionExtensions`.

This allows us to remove `Bond.BondReflection` during or after the next
major version of C# Bond.

Fixes https://github.com/Microsoft/bond/issues/369
