+++ cs/src/core/Transcoder.cs
@@ -9,7 +9,7 @@ namespace Bond
     using Bond.Expressions;
     using Bond.IO;
     using Bond.Protocols;
-    using Bond.Reflection;
+    using Bond.Internal.Reflection;
 
     /// <summary>
     /// Transcode payload from one protocol into another
