+++ cs/src/core/expressions/SerializerGeneratorFactory.cs
@@ -6,7 +6,7 @@ namespace Bond.Expressions
     using System;
     using System.Globalization;
     using System.Linq.Expressions;
-    using Bond.Reflection;
+    using Bond.Internal.Reflection;
 
     internal static class SerializerGeneratorFactory<R, W>
     {
