+++ cs/src/core/expressions/UntaggedParser.cs
@@ -8,7 +8,7 @@ namespace Bond.Expressions
     using System.Diagnostics;
     using System.Linq;
     using System.Linq.Expressions;
-    using Bond.Reflection;
+    using Bond.Internal.Reflection;
 
     public class UntaggedParser<R> : IParser
     {
