+++ cs/test/core/BondedTests.cs
@@ -7,7 +7,7 @@
     using Bond.Protocols;
     using Bond.IO;
     using Bond.IO.Unsafe;
-    using Bond.Reflection;
+    using Bond.Internal.Reflection;
 
     [TestFixture]
     public class BondedTests
