+++ cs/test/core/SerializerGeneratorFactoryTests.cs
@@ -8,7 +8,7 @@
     using Bond.Expressions;
     using Bond.Protocols;
     using Bond.IO.Unsafe;
-    using Bond.Reflection;
+    using Bond.Internal.Reflection;
 
     [TestFixture]
     public class SerializerGeneratorFactoryTests
