+++ cs/test/core/Util.cs
@@ -996,7 +996,7 @@ namespace UnitTest
                 streamTranscode(SerializeSP, TranscodeSPFB<From>, DeserializeFB<To>);
 
                 // Pull parser doesn't support bonded<T>
-                if (AnyField<From>(BondReflection.IsBonded))
+                if (AnyField<From>(Reflection.IsBonded))
                 {
                     streamTranscode(SerializeSP, TranscodeSPXml<From>, DeserializeXml<To>);
                 
@@ -1018,7 +1018,7 @@ namespace UnitTest
             }
 
             // Pull parser doesn't supprot bonded<T>
-            if (AnyField<From>(BondReflection.IsBonded))
+            if (AnyField<From>(Reflection.IsBonded))
             {
                 streamRoundtrip(SerializeXml, DeserializeXml<To>);
                 streamTranscode(SerializeCB, TranscodeCBXml<From>, DeserializeXml<To>);
